backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
huey[postgres]==2.5.3
psycopg2-binary==2.9.10
python-multipart==0.0.20
yt-dlp==2025.5.22
alembic==1.16.1
sqlalchemy==2.0.41
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0 